update travis not to treat warnings as failing build
The R CMD CHECK always gives a warning if some functions are not documented. This is almost always the case. Hence warnings shouldn't make the build fail. It would be even better to be able to be more specific because there might be valid warnings which should result in a build fail. Unfortunately I didn't find a way to be that specific so it's disabled for all warnings in the .travis.yml file.