add script to tidy
Have to load some packages separately so make sure those are installed. Functions are explicitly accessed in their namespace by using ::. The path is hardcoded but should be made a command line argument. Could have created separate functions in reproseries just to write to csv but decided this is overkill.